refactor

+++ src/services/user.service.ts
@@ -1,7 +1,10 @@
-import { UserRepository } from "../repositories/user.repository";
-import { User } from "../entities/user.entity";
 import bcrypt from "bcryptjs";
+import { UserRepository } from "../repositories/user.repository";
+import { User, UserRole } from "../entities/user.entity";
+import ForbiddenError from "../exceptions/forbidden-exception";
+import UserAlreadyExistsException from "../exceptions/user-exists-exception";
 
+// TODO types (no any)
 export class UserService {
   private userRepository: UserRepository;
 
@@ -12,25 +15,53 @@ export class UserService {
   async create(user: User): Promise<User> {
     const existingUser = await this.userRepository.findByEmail(user.email);
     if (existingUser) {
-      throw new Error("Email already in use"); // TODO: imrove error handling
+      throw new UserAlreadyExistsException(existingUser.email);
     }
     user.password = await bcrypt.hash(user.password, 10);
     return this.userRepository.create(user);
   }
 
-  async findAll(): Promise<User[]> {
+  async findAll(user: any): Promise<User[]> {
+    const isUser = user?.role === UserRole.USER;
+
+    if (isUser) {
+      throw new ForbiddenError();
+    }
+
     return this.userRepository.findAll();
   }
 
-  async findById(id: number): Promise<User | undefined> {
+  async findById(user: any, id: number): Promise<User | undefined> {
+    const isUser = user?.role === UserRole.USER;
+
+    if (isUser && user?.id !== id) {
+      throw new ForbiddenError();
+    }
+
     return this.userRepository.findById(id);
   }
 
-  async update(id: number, user: Partial<User>): Promise<void> {
-    await this.userRepository.update(id, user);
+  async update(id: number, user: any, data: Partial<User>): Promise<void> {
+    const isUser = user?.role === UserRole.USER;
+
+    if (isUser && user?.id !== id) {
+      throw new ForbiddenError();
+    }
+
+    if (data.password) {
+      data.password = await bcrypt.hash(data.password, 10);
+    }
+
+    await this.userRepository.update(id, data);
   }
 
-  async deleteById(id: number): Promise<void> {
+  async deleteById(id: number, user: any): Promise<void> {
+    const isUser = user?.role === UserRole.USER;
+
+    if (isUser || user?.id === id) {
+      throw new ForbiddenError();
+    }
+
     await this.userRepository.deleteById(id);
   }
 }
